сортирует файлы
