+++ src/const.js
@@ -3,3 +3,10 @@ export const SortType = {
   DATE: 'date',
   RATING: 'rating',
 };
+
+export const RenderPosition = {
+  BEFOREBEGIN: 'beforebegin',
+  AFTERBEGIN: 'afterbegin',
+  BEFOREEND: 'beforeend',
+  AFTEREND: 'afterend',
+};
